Fix: wrap useSearchParams in Suspense boundary for Next.js 15 compatibility
- Split page into server component with Suspense and client component
- Moved useSearchParams logic to separate UnsubscribeClient component
- Fixes 'useSearchParams() should be wrapped in a suspense boundary' error
- Maintains all functionality while following Next.js 15 requirements
